moved updating heatmap collection to pipeline utils and added updating heatmap collection in ingestion removal pipeline
